rename ssh.py to avoid module conflict with ssh package in new ncclient version
